fix compilation warnings
git-svn-id: http://svn2.assembla.com/svn/secu3man@225 c1dae7de-a428-49f4-85d8-50b65f41d01d

+++ sources/secu3man/TabDialogs/FirmwareTabDlg.cpp
@@ -265,7 +265,7 @@ CFirmwareTabDlg::CFirmwareTabDlg(CWnd* pParent /*=NULL*/)
  int rpm = 200;
  for(size_t i = 0; i < 128; i++)
  {
-  m_attenuator_table_slots[i] = rpm;
+  m_attenuator_table_slots[i] = (float)rpm;
   rpm+=60;
  }
 }
